pyproject.toml
[project]
name = "cat12-bids-pipeline"
version = "1.0.0"
description = "BIDS-compatible pipeline for longitudinal neuroimaging data processing using CAT12 standalone"
authors = [
    {name = "MRI Lab Graz", email = "contact@mri-lab-graz.org"},
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
keywords = ["neuroimaging", "bids", "cat12", "longitudinal", "mri"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    # Core BIDS and neuroimaging libraries
    "nibabel>=3.2.1",
    "pybids>=0.15.1",
    "pandas>=1.3.0",
    "numpy>=1.21.0",
    
    # File and data manipulation
    "pyyaml>=6.0",
    "click>=8.0.0",
    "tqdm>=4.62.0",
    
    # Logging and utilities
    "colorlog>=6.6.0",
    "jinja2>=3.0.0",
    
    # Quality assurance and validation
    "bids-validator>=1.9.0",
    
    # Optional: Parallel processing
    "joblib>=1.1.0",
]

[project.optional-dependencies]
visualization = [
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=5.0.0",
    "mypy>=0.991",
]
all = [
    "cat12-bids-pipeline[visualization,dev]",
]

[project.urls]
Homepage = "https://github.com/MRI-Lab-Graz/cat-12"
Documentation = "https://github.com/MRI-Lab-Graz/cat-12#readme"
Repository = "https://github.com/MRI-Lab-Graz/cat-12.git"
"Bug Tracker" = "https://github.com/MRI-Lab-Graz/cat-12/issues"

[project.scripts]
cat12-bids-processor = "bids_cat12_processor:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["utils", "scripts"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | external
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["utils", "scripts"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
exclude = [
    "external/",
    ".venv/",
    "build/",
    "dist/",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=utils",
    "--cov=scripts",
    "--cov-report=html",
    "--cov-report=term-missing",
]

[tool.coverage.run]
source = ["utils", "scripts"]
omit = [
    "*/tests/*",
    "*/external/*",
    "*/.venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]